+++ ember-flight-icons/tests/dummy/app/components/ds-icon-grid.hbs
@@ -16,7 +16,7 @@
         placeholder="e.g. arrow"
         autofocus={{false}}
         value={{this.search}}
-        {{on "input" this.debouncedUpdate}}
+        {{on "input" this.args.onUpdate}}
       />
 
     </div>
@@ -44,7 +44,7 @@
   </div>
 </div>
 <ul class="ds-ul-grid" data-test-target="icon-grid">
-  {{#each this.icons as |meta|}}
+  {{#each @icons as |meta|}}
     <li class="ds-li {{if meta.isHidden "d-none"}} {{if (not-eq meta.size this.visibleIconSize) "d-none"}}">
       <div class="ds-icon-frame">
         <FlightIcon @name={{meta.iconName}} @size="{{meta.size}}" class="demo-icon" />
