+++ ember-flight-icons/tests/dummy/app/routes/index.js
@@ -1,3 +1,20 @@
 import Route from '@ember/routing/route';
 
-export default class IndexRoute extends Route {}
+import catalog from '@hashicorp/flight-icons/catalog.json';
+export default class IndexRoute extends Route {
+  model() {
+    this.icons = catalog.assets.map(
+      ({ iconName, fileName, size, description }) => {
+        return {
+          iconName: `${iconName}`,
+          name: `${fileName}`,
+          size: `${size}`,
+          description: `${description}`,
+          searchable: `${iconName}, ${description}`,
+        };
+      }
+    );
+
+    return this.icons;
+  }
+}
