+++ ember-flight-icons/tests/dummy/app/templates/index.hbs
@@ -54,5 +54,5 @@
 
 <section class="ds-section" aria-labelledby="available-icons">
   <h2 class="ds-h2 mb-3" id="available-icons">Available Icons</h2>
-  <DsIconGrid />
+  <DsIconGrid @searchQuery={{this.query}} @icons={{this.filteredIcons}} @onUpdate={{this.debouncedUpdate}} />
 </section>
